Add files via upload
Exercise for practising with regularised linear regression. We use a dataset of water flowing out a dam against change in the water level. We try to fit a polynomial of different degrees to practice with concepts as underfitting or overfitting.  Worked with bias-variance tradeoff to find the best value for the regularising parameter.